refactor(carousel): migrate v2 to main and remove legacy code
