+++ components/shared/Carousel/Carousel.type.ts
