+++ components/shared/Carousel/index.tsx
@@ -2,6 +2,6 @@ import Carousel from "./Carousel";
 
 export * from "./Carousel";
 
-export * from "./mockData";
+export * from "./Carousel.type";
 
 export default Carousel;
