--- components/shared/Carousel/v2/index.tsx
@@ -1,7 +1,0 @@
-import Carousel from "./Carousel";
-
-export * from "./Carousel";
-
-export * from "./Carousel.type";
-
-export default Carousel;
